Extract + configure an embed viewer the blacklight way
