+++ app/components/arclight/embed_component.html.erb
@@ -1,0 +1,13 @@
+<%= tag.p class: "media breadcrumb-item breadcrumb-item-#{depth + 3}" do %>
+  <span class="media-body al-online-content-icon" aria-hidden="true"><%= helpers.blacklight_icon :online %></span>
+  <span class="col text-muted">
+    <%= t('arclight.views.show.online_content_upper') %>
+  </span>
+  <%= render Arclight::OembedViewerComponent.with_collection(embeddable_resources, document: @document, depth: depth + 6) %>
+
+  <% linked_resources.each do |resource| %>
+    <div class="al-digital-object media breadcrumb-item breadcrumb-item-<%= depth + 6 %>">
+      <%= link_to(resource.label, resource.href) %>
+    </div>
+  <% end %>
+<% end %>
