+++ lib/arclight/engine.rb
@@ -8,14 +8,12 @@ require 'arclight/normalized_date'
 require 'arclight/normalized_id'
 require 'arclight/normalized_title'
 require 'arclight/digital_object'
-require 'arclight/viewer'
 require 'blacklight_range_limit'
 
 module Arclight
   ##
   # This is the defining class for the Arclight Rails Engine
   class Engine < ::Rails::Engine
-    config.viewer_class = Arclight::Viewers::OEmbed
     config.oembed_resource_exclude_patterns = [/\.pdf$/, /\.ppt$/]
 
     Arclight::Engine.config.catalog_controller_group_query_params = {
