+++ lib/generators/arclight/templates/catalog_controller.rb
@@ -66,6 +66,7 @@ class CatalogController < ApplicationController
     # solr field configuration for document/show views
     # config.show.title_field = 'title_display'
     config.show.document_component = Arclight::DocumentComponent
+    config.show.embed_component = Arclight::EmbedComponent
     config.show.display_type_field = 'level_ssm'
     # config.show.thumbnail_field = 'thumbnail_path_ss'
     config.show.document_presenter_class = Arclight::ShowPresenter
